+++ test/utils.js
@@ -5,7 +5,7 @@ import nock from 'nock';
 
 module.exports = {
   createNock() {
-    return nock('http://ps.pndsn.com:80', {
+    return nock('http://pn.pndsn.com:80', {
       filteringScope: () => true
     });
   },